Lots of minor type fixing, standardize MPL header format
