+++ src/client/app/containers/LineChartContainer.ts
@@ -1,8 +1,6 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
+/* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 import * as _ from 'lodash';
 import { Line, ChartComponentProps } from 'react-chartjs-2';
@@ -12,9 +10,7 @@ import { connect } from 'react-redux';
 import getGraphColor from '../utils/getGraphColor';
 import { State } from '../types/redux/state';
 
-/**
- * @param {State} state
- */
+
 function mapStateToProps(state: State) {
 	const timeInterval = state.graph.timeInterval;
 	const datasets: ChartDataSets[] = [];
