+++ src/client/app/containers/groups/GroupSidebarContainer.tsx
@@ -4,7 +4,8 @@
 
 import * as _ from 'lodash';
 import { connect } from 'react-redux';
-import { changeDisplayedGroups, changeDisplayMode, DisplayMode } from '../../actions/groups';
+import { changeDisplayedGroups, changeDisplayMode } from '../../actions/groups';
+import { DisplayMode } from '../../types/redux/groups';
 import GroupSidebarComponent from '../../components/groups/GroupSidebarComponent';
 import { Dispatch } from '../../types/redux/actions';
 import { State } from '../../types/redux/state';
