+++ src/client/app/reducers/barReadings.ts
@@ -1,9 +1,6 @@
-
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
+/* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 import { BarReadingsAction, BarReadingsState } from '../types/redux/barReadings';
 import { ActionType } from '../types/redux/actions';
