+++ src/client/app/reducers/notifications.ts
@@ -1,10 +1,7 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
+/* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { Notification } from 'react-notification-system';
 import { NotificationAction, NotificationsState } from '../types/redux/notifications';
 import { ActionType } from '../types/redux/actions';
