+++ src/client/app/types/redux/admin.ts
@@ -1,0 +1,55 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+import {PreferenceRequestItem} from '../items';
+import { ChartTypes } from './graph';
+import { ActionType } from './actions';
+
+export type AdminAction =
+	| UpdateDisplayTitleAction
+	| UpdateDefaultChartToRenderAction
+	| ToggleDefaultBarStackingAction
+	| RequestPreferencesAction
+	| ReceivePreferencesAction
+	| MarkPreferencesNotSubmittedAction
+	| MarkPreferencesSubmittedAction;
+
+export interface UpdateDisplayTitleAction {
+	type: ActionType.UpdateDisplayTitle;
+	displayTitle: string;
+}
+
+export interface UpdateDefaultChartToRenderAction {
+	type: ActionType.UpdateDefaultChartToRender;
+	defaultChartToRender: ChartTypes;
+}
+
+export interface ToggleDefaultBarStackingAction {
+	type: ActionType.ToggleDefaultBarStacking;
+}
+
+export interface RequestPreferencesAction {
+	type: ActionType.RequestPreferences;
+}
+
+export interface ReceivePreferencesAction {
+	type: ActionType.ReceivePreferences;
+	data: PreferenceRequestItem;
+}
+
+export interface MarkPreferencesNotSubmittedAction {
+	type: ActionType.MarkPreferencesNotSubmitted;
+}
+
+export interface MarkPreferencesSubmittedAction {
+	type: ActionType.MarkPreferencesSubmitted;
+}
+
+export interface AdminState {
+	displayTitle: string;
+	defaultChartToRender: ChartTypes;
+	defaultBarStacking: boolean;
+	isFetching: boolean;
+	submitted: boolean;
+}
