+++ src/client/app/types/redux/lineReadings.ts
@@ -1,11 +1,7 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
+/* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import axios from 'axios';
-import * as moment from 'moment';
 import { TimeInterval } from '../../../../common/TimeInterval';
 import { LineReadings } from '../readings';
 import { ActionType } from './actions';
